requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
supabase==2.0.2
pydantic==2.5.0
pydantic-settings==2.1.0
httpx>=0.24.0,<0.25.0
redis==5.0.1
celery==5.3.4
mysql-connector-python==8.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
gunicorn==21.2.0 